Basic functionality complete
At this state memory of any size can be allocated and accessed, but not
freed yet. Memory is accessed in a slightly roundabout way in order to
be able to totally release memory even if other allocated memory is on
top of it, this functionality will be included when the ability to free
memory is included.