Split retrive_memory() into two functions
Sometimes we need the address of the memory meta block, but other times
we need the address of the memory itself. Two different functions will
handle this now, with retrieve_memory_meta_block() returning the address
of the memory meta block, whereas retrieve_memory() will call
retrieve_memory_meta_block() and add one to it(thus pointing to the
memory itself) before returning that value.