Convert pointer based linked-list into + based.
Since the potential exists for the entire heap to move when a memory
block is freed, the pointers in mem_block_meta are not guarenteed to
point to the next mem_block_meta. However, since the meta blocks contain
the size of the succeeding memory block, simple pointer arithmetic can
be used to find the next meta block, in a wholly relative way.
Therefore, this new memory traversal method is agnostic to the absolute
position of the heap as a whole, making it far more flexible.

However, this method does sacrifice a little speed since arithmetic and
sizeof evalutations need to be done each time instead of a simple
dereference.